refactor: move to own class
